% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyconfig.R
\docType{package}
\name{shinyconfig}
\alias{shinyconfig}
\alias{shinyconfig-package}
\title{shinyconfig: A package for computating the notorious bar statistic.}
\description{
The shinyconfig package provides functions to aid in the configuration and setup for shiny apps as standalone apps or packages.
}
\section{shinyconfig functions}{

}

